pkg/flow: move scheduler to the internal controller package
